+++ packages/shadcn-formily/src/components/combobox.tsx
@@ -1,10 +1,11 @@
 'use client';
 
+import type { FC } from 'react';
 import { connect, mapProps } from '@formily/react';
 
 import { Combobox as BaseCombobox } from '@pixpilot/shadcn-ui';
 
-export const Combobox = connect(
+export const Combobox: FC = connect(
   BaseCombobox,
   mapProps({
     dataSource: 'options',