+++ packages/shadcn-formily/src/components/select.tsx
@@ -1,8 +1,9 @@
+import type { FC } from 'react';
 import { connect, mapProps } from '@formily/react';
 
 import { Select as BaseSelect } from '@pixpilot/shadcn-ui';
 
-export const Select = connect(
+export const Select: FC = connect(
   BaseSelect,
   mapProps({
     dataSource: 'options',