+++ packages/shadcn-ui/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @pixpilot/shadcn-ui
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @pixpilot/shadcn@0.1.11
+
+## 0.0.6
+
+### Patch Changes
+
+- fix release
+- Updated dependencies
+  - @pixpilot/shadcn@0.1.10
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @pixpilot/shadcn@0.1.9
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @pixpilot/shadcn@0.1.8
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @pixpilot/shadcn@0.1.7
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @pixpilot/shadcn@0.1.6
+
 ## 0.0.1
 
 ### Patch Changes