+++ packages/shadcn/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @pixpilot/shadcn
 
+## 0.1.11
+
+### Patch Changes
+
+- improve button
+
+## 0.1.10
+
+### Patch Changes
+
+- fix release
+
+## 0.1.9
+
+### Patch Changes
+
+- fix bundler
+
+## 0.1.8
+
+### Patch Changes
+
+- test
+
+## 0.1.7
+
+### Patch Changes
+
+- test publish
+
+## 0.1.6
+
+### Patch Changes
+
+- test publish
+
 ## 0.1.0
 
 ### Minor Changes