+++ tooling/storybook/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @internal/storybook
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @pixpilot/shadcn@0.1.11
+  - @pixpilot/shadcn-auth@0.0.7
+  - @pixpilot/shadcn-formily@0.0.7
+  - @pixpilot/shadcn-ui@0.0.7
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @pixpilot/shadcn-formily@0.0.6
+  - @pixpilot/shadcn-auth@0.0.6
+  - @pixpilot/shadcn-ui@0.0.6
+  - @pixpilot/shadcn@0.1.10
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @pixpilot/shadcn@0.1.9
+  - @pixpilot/shadcn-auth@0.0.5
+  - @pixpilot/shadcn-formily@0.0.5
+  - @pixpilot/shadcn-ui@0.0.5
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @pixpilot/shadcn@0.1.8
+  - @pixpilot/shadcn-auth@0.0.4
+  - @pixpilot/shadcn-formily@0.0.4
+  - @pixpilot/shadcn-ui@0.0.4
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @pixpilot/shadcn@0.1.7
+  - @pixpilot/shadcn-auth@0.0.3
+  - @pixpilot/shadcn-formily@0.0.3
+  - @pixpilot/shadcn-ui@0.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @pixpilot/shadcn@0.1.6
+  - @pixpilot/shadcn-auth@0.0.2
+  - @pixpilot/shadcn-formily@0.0.2
+  - @pixpilot/shadcn-ui@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes